Receive plugins' versions with underscores. (#699)
* Receive plugins versions with underscores.

New jenkins plugin versions have underscores in them (example: https://plugins.jenkins.io/credentials/). The operator refuses configuration for such plug-ins as they do not match the existing regex.

* Add tests to underscores in plugin versions

Adds a new test to check that versions with underscores are valid.

* Synchronise regexps character order to avoid range

Make two related regexps more evidently similar to each other and re-order the symbols so that the - is not interpreted.